HA-62: Relocate Constant Collection in Presentation WebAPI
